+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in v1.4.3
   to Get-Printer-Attributes responses.
 - Fixed missing mutex unlock in DNS-SD code (Issue #299)
 - Fixed "printer-id" value for new printers (Issue #301)
+- Fixed DNS-SD device list crash (Issue #302)
 - Fixed default "copies" value with `papplJobCreateWithFile`.